[OM-65319]: Support ConsistentScaling on heterogeneous providers RB:43199
** Problem **

We model a group of entities replicated from some shared definition using what we call a “Consistent Scaling Group”. Because they are replicas of each other, members of a consistent scaling group must share certain features in common. The most important of these is that commodities resized across the scaling group must share the same capacity.

We have problems with consistent scaling CPU resources when the suppliers for the scaling group members have different CPU speeds. CPU resource values (capacity, used, etc.) are all normalized to be of the same unit (normalized MHz). CPU resources for scaling groups are usually defined in terms of cores (for VMs) or millicores (for containers) and not in MHz (and certainly not Turbo normalized MHz). CPU resource capacities for a scaling group are required to be consistent in this original group unit of cores/millicores. If all group members are running on providers with the same Turbo normalized CPU speeds, there is no problem - the originally consistent capacities in cores/millicores will be consistent in the normalized units as well. However, if, for example, some group members are running on providers with fast CPU speeds and some members are running on providers with slow CPU speeds, because the market sees capacities in normalized units, the market sees inconsistent capacities for the group members even though they are consistent in their original unit.

The market responds to this inconsistency by trying to make all group members consistent. It picks the highest capacity from all group members and generates a resize action to take all group members to this capacity. In our example above, this would mean the market would generate a resize action to scale the group from 200MHz to 300MHz in order to make all group members consistent. Unfortunately, this is a nonsense action because the capacities were actually already consistent, just not in the units the market was using. A similar situation happened at a customer where for example where we resized up a container at 2% utilization because of this inconsistency.

** Solution **

We introduce a new double value (called "consistentScalingFactor") on TraderSettings that can be used to convert from the normalized economy units (ie MHz) where group members might not be consistent back to group-local units (ie millicores/cores) where group members should be consistent. Right now the new consistentScalingFactor is only applied in the ConsistentResizer class responsible for reconciling the individual PartialResize actions for the individuals in the scaling group which leaves a small correctness gap around Container quota commodities to be addressed later. To ensure that the CSF is applied properly, we force conversion of normal market quantities into ConsistentScalingQuantities. Because ConsistentScalingQuantities can only do arithmetic and comparisons with other ConsistentScalingQuantities, this makes it hard to forget to convert a quantity in normalized units into consistent units before doing arithmetic/comparisons with a quantity already converted to consistent units. At the very end, we convert back from consistent units into normalized units so that the rest of the market after the ConsistentResizer only has to deal with the normalized units as it does prior to this change.

Note that because the CSF is introduced on TraderSettings rather than on CommoditySoldSettings, it would naturally be applied to all commodities on that trader. However, we only want it applied to some commodities (currently the CPU commodities). To prevent the CSF from being incorrectly applied to all commodities, we update the RawMaterials to include a flag indicating whether a particular commodity should apply the ConsistentScalingFactor or not.

To account for a trader that may have moved before we generate a resize, we use the consistentScalingFactor from the selling trader's provider when converting the resize's newCapacity into consistent units. Note that the oldCapacity and all settings should use the consistentScalingFactor from the starting provider for the selling trader. RawMaterial commodity values should use the CSF of the trader selling that raw material.

** Testing Done **

* Duplicated all ConsistentResizer tests to also test behavior in a heterogeneous provider scenario (where members of a consistent scaling group have different ConsistentScalingFactors).
* Ensured all market tests still pass.
* New unit tests for ConsistentScalingQuantity class.
* Tested in-app to ensure that when containers in a consistent scaling group are spread across a set of heterogeneous providers, we can still get a sensible resize action.
* Executed an action in the lab on a heterogeneous cluster (dc11 target) and saw the deployment correctly updated.
* Verified VMs that are not part of the kubernetes cluster still get the same resize actions.