[OM-62376] Initial fix for the log pollution in EntityValidator.
Added code to count the number of times a given log message is printed
and suppress messages after a fixed limit is exceeded. Also added code
to print the number and pattern of suppressed messages.

Tested code using Voltron by adding a Nutanix target known to cause
these particular warnings to be printed and verifying only the first n
messages are printed.

Added some instrumentation code to get a feel of the overhead added to
the logging code. It appears to be less than 10%. i.e. log messages take
tens of microseconds to print while figuring out whether to print them
takes microseconds. This code will be removed in a later commit.

The corresponding messages for bought commodities were moved back to
warning level and need to use a similar mechanism to avoid clutter.

The fastutil library was used to reduce memory footprint.