[OM-64212]: XL Reporting: Split actions into pending and completed.
Intent
======
Create a cleaner and more flexible representation for actions in reporting.

Background
==========
Pending and historical actions are fundamentally different, and we want the reporting tables to represent and support that. Historical actions need long retention periods, and near-realtime ingestion. Pending actions represent the current snapshot of the action orchestrator's data. The current storage model mixes pending and historical together, which makes it harder to treat them differently. Furthermore, it is confusing, and the difference between an "action spec" and an "action" is artificial and not useful for reporting purposes.

The new model has two tables - **pending_action** for currently recommended actions, and **completed_action** for completed actions. **completed_action** is a hypertable with 7-day uncompressed chunks. We replace the contents of **pending_action** with the latest data from the action orchestrator at a configurable interval (every 1 hr by default). We write to the **completed_action** table whenever actions succeed.

Implementation
==============
Action Orchestrator
===================
I changed the action orchestrator to send the translated ActionSpecs together with the ActionSuccess and ActionFailure notifications. This makes it a lot easier for the receiver, because it eliminates a range of error conditions (AO down or unavailable, action not found) and speeds up processing. Each action spec is fairly small, so there is virtually no cost to doing this.

Extractor
=========
- The pending action code is cherry-picked from https://rbcommons.com/s/VMTurbo/r/42958/. The only change is adding the recommendation_time, so we have an idea of how long a particular pending action has been recommended.
- **CompletedActionWriter** contains all the logic for writing the completed data.
    - Completed actions get converted to database records and put in a queue.
    - A background thread (**RecordBatchWriter**) picks up batches of actions from the queue and writes them to the database using upserts.
        - The reason we use a background thread is to handle bursts of completed actions. 1000 completed actions will generate 1000 separate notifications, but we don't want to do 1000 separate inserts if we don't have to.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/43091