[OM-65194]: Ignore insertion failure  when restoring settings and policies from diagnostics
Intent
======
Prevent diagnostics restoration failure if some policies have missing groups as a result of race condition while exporting diagnostics

Problem
=======
Diagnostics restoration failed while restoring JBHunt and BofA diagnostics. The error indicates that some groups are missing for some of the setting policies.

Diagnosis
=========
The group information gets exported before the setting and policy information. As group component is not frozen in the time of exporting diagnostics, new setting and groups can be added while we are exporting the diagnostics. This may result in the fact that the setting policy gets exported but its group is not in diagnostics.

Solution
========
Two different solution considered for this issue:
- Freeze the group component interactions while exporting the diagnostics. This prevents the race condition and assures that all the groups for setting policies are there.
- Dont give up importing diagnostic when a failure occurrs in importing diagnostics
After talking with Andy, Roman and Joseph about this options, we came up with the second solution is better as first one may have effect on customer experience.
This solution adds ignore to table insertions commands while restoring diagnostics so if a insert fails the diagnostics restoration continues.