[OM-62376] Created a new logger class and used it in EntityValidator.
The new logger implements log4j's `Logger` interface while
encapsulating an existing logger to which it delegates most calls
(decorator pattern). The result is that the new object has all the
convenience methods of the loggers we are used to using and can be used
as a drop-in replacement to fix the log pollution problems of an entire
class.

Tried 3 or 4 different ways to implement the class using different
customisation points provided by the log4j class hierarchy. I avoided
overriding implicitly pure methods to be impure for obvious reasons, but
had to choose between overriding a log method that is supposed to always
log something with one that might not and overriding more than a dozen
methods that weren't intended to be used as customisation points with
repetitive code and I chose the former. Given the design of the logger
classes it seemed more likely that a new method would be added that we
wouldn't automatically know to override than not printing causing a
problem. Also less boilerplate code would make maintenance easier.

Also wrote a small unit test and tested manually with Voltron and a
Nutanix target from our lab. Made the threshold 1 for testing but will
increase it in a later commit.