[OM-65750]: XL Action: Send SUCCESS/FAILURE notifications only when actions are in terminal states.
Issue
=====
If an action fails and post-script fails we send two action failure notifications.

Background
==========
The **ActionStateUpdater** receives success and failure notifications from the topology processor and manipulates the state machine of the action accordingly. For actions with pre and post scripts there can be multiple invocations of "success" and "failure" for a single action ID. When an action reaches a terminal state (SUCCEEDED or FAILED) the action orchestrator should send out a notification. Other components can operate on the action - for example, the extractor component writes it to Postgres for reports.

The issue is that we are sending out failure notifications when the action transitions to **FAILING**, which is the non-terminal state for an action that fails, but has a POST script that still needs to run. There are two problems:
1) The action is not yet in a terminal state at the time the notification goes out, which goes against the (implied) notifications interface.
2) If the POST script also fails, the **ActionStateUpdater#onActionFailure** method will get called again (for the POST script) and we will send out a second failure notification. Again, this is unexpected.

Fix
===
The fix is twofold:
1) In ActionStateUpdater#onActionFailure, only send out a failure notification if the action has transitioned to the FAILED state.
2) In ActionStateUpdater#onActionSuccess, send out a failure notification if the action has transitioned to a FAILED state. This happens when a POST script of a failed action succeeds.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/43170/