[OM-62376] Moved new logger from xl to opsmgr repository to be shared.
The intention is for this new implementation of the log4j `Logger`
interface to be used by all components and code in both repositories to
help reduce log pollution. Currently the best project to place it
appears to be the `com.vmturbo.commons` one in `opsmgr` repository.

Also moved the existing unit test there and created a new log4j test
configuration for that project as it didn't have one and the test was
breaking.

Updated poms accordingly to add the fastutils dependency to the
`com.vmturbo.commons` project and hoist common configuration for that
project to parent pom.