[OM-62376] Moved new logger from xl to opsmgr repository to be shared.
The intention is for this new implementation of the log4j `Logger`
interface to be used by all components and code in both repositories to
help reduce log pollution. Currently the best project to place it
appears to be the `com.vmturbo.commons` one in `opsmgr` repository.

Also moved the existing unit test there and removed a duplicate
fastutils dependency from parent xl pom.

Run the unit test and also retested with Voltron and the Nutanix target
in our lab and code appears to still work.