[OM-64941] RB-43132 Add timeout to free the executor used to export diags in api
Intent:

Alleviate a bug

Background:

In order for diagnostics to be exported a POST request to /admin/exports has to be made. The api component will submit
a new task to a SingleThreadExecutor that will use the clustermgr API to invoke the export in the clustermgr component.

It was reported that the api component was not responding to requests for exporting diagnostics. There were no logs
produced in api and clustermgr. Triggering the export directly in the clustermgr worked as expected. Furthermore,
restarting the api pod solved the problem on the customer side.

Based on all the above it seems that the thread submited to the executor can get stuck and there is no fail-safe
mechanism to ensure termination.

IMPORTANT: I have not been able to reproduce the problem, so this is a proposed safety-net approach.

NOT A FIX:

A timeout is now enforced to make sure the executor is freed up even if the worker thread is blocked. The timeout is set
to 6 hours by default since this is a safe period of time for the export to complete.

If the worker thread is blocking at a point where Thread.interrupted() is checked, then it will terminate and the
notification will be lost. This does not however seem to be the case with REST requests. The worker thread WILL NOT
terminate while waiting for ClusterMgrApi response. The executor is available to receive new requests, but the thread is
still running. This does not affect user experience. It is unsure if WebSocketSession#sendMessage() is blocking.
If it is - and the current data seem to support it - then it is unknown if it is aware of interrupts.

To sum up, the timeout enforced does not guarantee worker thread termination. It only frees up the SingleThreadExecutor
so that new requests can be submitted. This will mitigate the current issue, by unblocking the client side of exporting diags.

Squashed commit of the following:

commit ef5b14dcea158c69fa488a8549b6f4e909d301f5
Author: Alexandros Ioannou <alexandros.ioannou@turbonomic.com>
Date:   Mon Jan 18 16:56:19 2021 +0200

    Add Gary's comments

commit 1ce8f1e4326b94c9e66eb7087b5ce60ab946ec62
Author: Alexandros Ioannou <alexandros.ioannou@turbonomic.com>
Date:   Mon Jan 18 15:38:29 2021 +0200

    Add Dmitry's comments

commit 0af3fd5a8dafbd6c5fe1a09afc8003ca7d259ee5
Author: Alexandros Ioannou <alexandros.ioannou@turbonomic.com>
Date:   Fri Jan 15 14:08:15 2021 +0200

    Add timeout to api