[OM-61842] RB-43277 Unblock delete of multiple licenses
Intent:

Fix a bug.

Background:

When deleting multiple licenses we validate the Uuids by calling the License GRPC Service.
This step was blocking, causing the api to never produce a response. The blocking was traced
to the server-side of the communication (auth component).

Fix:

The problem was caused because the server-side was never calling onCompleted() to finalize
the response, so the connection was hunging, and the client was waiting for more information
to come in. By adding this call the problem was fixed.