[OM-62376] Added JUnitParams as a dependency of com.vmturbo.commons.
That was done in anticipation of using it to extend the test suite of
the new logger.

Since the artefact is used by more than one module, the common
attributes of the dependency were hoisted to the parent pom. The JUnit
dependency was also updated to take advantage of the dependency
management section of the parent pom.

Also moved some dependencies around in the com.vmturbo.commons pom so
that only test-scope dependencies are under the `<!-- Test Scope -->`
comment.

Tried to import a JUnit class from production code to make sure I didn't
accidentally make it available there.

Note that the various POM files are in need of further cleanup and
deduplication.