[OM-63732]: XL Group - Pagination support for groups stage 2: additional filtering
The second part of the group pagination effort in group component.
Added support to filter by:
- non-empty groups
- group environment & cloud type
- static/dynamic groups
- member type
- entity type

Introduced storing of environment type & cloud type in group db,
as well as a small pipeline that runs after repository receives a new
topology, and updates entity-derived information for groups.

NOTE: none of these changes are currently visible to the users.
Integration with the api will be done in a later stage of this effort.

RB42765