[OM-66485]: Fixed the issue that prevented changing policy defaults in loaded diags
Intent
======
Fixed a mistake in a condition that caused issues when updating default policy in certian cases

Problem
=======
In loaded Credit Suisse diags, we see an error when trying to update the default policies for the vm.

Diagnosis
=========
If a setting for a policy is not sent from UI and it should have a value. We have a logic in place that goes with existing value if that exists or default value for that setting if that there is no existing logic. I made a mistake and used variable in the check for checking if we should go with default value or existing value and we always go with existing value if does not exists which causes the issue.

Solution
========
I updated the condition to use correct variable.

Testing Done
============
- Tested with Credit Suisse Diag and the issue is no longer present
- I added a unit test which I should have add the first time to test the case

Code Coverage
=============
No significat change