[OM-65935] Topology Processor -  Support percentile calculation for IOThroughput commodities bought from Compute Tier.
Objective
---------
The goal of this Story is to add support for percentile calculation for IOThroughput commodity bought by VMs from Compute Tiers. This will enable M2 to calculate the IOThroughput desired demand based on the percentile utilization and the resize target utilization setting. Consequently, VM Scale Actions will be based on IOThroughput percentile performance in addition to the existing commodities.
Note that this change enables IOThroughput percentile calculation in Topology Processor. However, the Azure Probe must create the CommodityDTOs for IOThroughput with the utilizationData field set. This is done in a separate change.

Testing
-------
Unit Tests
Added unit test to assert that IOThroughput is an applicable commodity provided that utilizationData is present.

Environment
8.1.1 instance with Azure targets added

1. Queried percentile_blobs table and observed that IOThroughput counts are stored for VMs
2. Generated Throughput stress on Azure VM and observed a Scale VM action to move the VM to a Tier with a higher Throughput capacity. Observed that the Action explanation states Throughput congestion.