[OM-63946] Move certificates to a Kubernetes secret
Background
==========
There are some encryption keys in XL that are currently stored in persistent volumes. These would be more appropriately stored as Kubernetes secrets. We are trying to reduce our usage of persistent volumes in Kubernetes, in order to reduce our reliance on Gluster and also to set us up better for potential HA designs in the future.

The following microservices are using persistent volumes to store certificates:

* api
* auth
* topology processor

Solution
========

A feature flag, `enableAuthSecrets`, enables new logic that reads the encryption keys from pre-existing Kubernetes secrets, instead of creating them on-demand and storing them in the PV. The same feature flag also instructs the `t8c.sh` script to generate these encryption keys during initial installation for new environments. Migration of the keys for existing environments is straightforward, but is not automated as part of this review.

Additionally, the private keys and admin JWT that are generated are now stored in Consul instead of in the PV.