[OM-66468]: Include ContainerSpec entities in plan scope
** Problem **
Container plan generate the wrong resize recommendations for containers in plan results. The reason is that the plan scoping logic only includes entities in the topology that have buys/sells relationships to the plan scope seed entities. ContainerSpec entities only have an AggregatedBy/Aggregator connection to the Container entities in the plan. However, because containers are ephemeral they cannot retain their own percentile history which is instead retained on the related ContainerSpec entities. By removing these ContainerSpcs during scoping, this prevents the copy of the shared history from ContainerSpec onto Container. Further, it prevents the correct propoagation of the ConsistentScalingFactor onto Containers in order for correct scaling decisions for these entities.

** Solution **
Include ContainerSpecs that are aggregators of containers in the plan scope.

** Testing Done **
* New unit test.
* Verified that plan sizing decisions for containers that used to differ between plan and real time (with no changes from real-time in the plan) now match whereas they used to be quite different in many cases.
* Impact on plan performance should be minimal because ContainerSpec entities are not controllable.