[OM-66453] Validate search scope for empty group
Problem: Searching for Cloud Commitments within the Azure billing family scope was returning an "illegal scope" error, causing the RI/SP widget to load incorrectly.

Cause: Since Azure doesn't have savings plans, the billing family search had no results. In such a case, a call is made to the repository to verify that the scope is a valid entity. If no entity is found, an exception is thrown. Since a billing family is not an entity, this exception was thrown even though the scope was valid.

Solution: When verifying the scope for an empty result, check for groups as well as entities. If either is found, the scope is valid and no exception is thrown.