[OM-66388]: Add a migration script that removes the setting values with empty string
Intent
======
Fixed the issue that settings that are empty are considered non-empty

Problem
=======
The actions are tried to be executed for multiple times in the customer environment

Diagnosis
=========
In 7.22.4, we keep an empty string in DB when the value of a setting is empty. After changes in OM-60946, we started not saving empty values in DB. However existing values were still in the DB. Therefore, when we upgraded to 8.0.4, the default setting value for these setting were not considered empty. Therefore, when an action gets accepted and we consider that there is a before action workflow (as this is one of those settings that by default should be not set)(``Action::getExecutionState``) and we go to ``PRE_IN_PROGRESS`` state. When we try to execute the before execution workflow of the action however there is this other logic that considers empty string as no workflow (``Action::getWorkflow``) and executes the action regularly. When the action is successful, we go to state ``IN_PROGRESS``  that we try to execute the action again and that explains double execution.

Solution
========
- I added a migrations that removes settings value that have empty string. Currently, we don't have a setting that there is meaningful different between having no-value and empty string value, therefore, this migration should not have any negative side effect.
- Additionally, I removed one wrong logic API. There is a condition in API that is alway true and result in sending empty string for default value to API. I changed that logic to act it was intended.
- I also removed a logic which checks the workflow setting and if the value is empty considers there is no workflow. I removed that logic as we don't represent no workflow as an empty string.