[OM-65738]: Introduce entity uptime calculation
**Background**

Our current topology cost and action savings make an assumption that the associated workload will be up 100% of the time. The UI extends this assumption in converting to a monthly rate by multiplying by 730 (to get a cloud month). In this code change, we start to calculate entity uptime, based on CCA demand.