[OM-66761]: Fix a unit test that fails in different timezones
Problem
=======
A number of tests were failing in different timezones

Diagnosis
=========
There is this logic in the api conversion project that we convert a timestamp to an string in offseted ISO 8601 format. In the test, I compare the conversion result with the string value in EST offset which is wrong as we may have different offset in different timezones.

Solution
========
I changed the code that compare the timestamp value of the converted rather than string value so it works in different timezones