[OM-66088] - Cache the group parents for entities and use it in entity group service
Intent
======
Include dynamic group that are parent of an entity in group for entities response.

Features Implemented
====================
We already cache the group members in group component. This caching mechanism has an up to date list of groups and their children. I updated this logic to also keep the entities to group link as well. In details, I did the following:
- I added another map from entity id to set of group ids that contain that entity
- I update the regroup logic and other logics that update the group to entities relationship to also update this map.
- I added a flag which enables/disables caching entity to group mapping