[OM-66854]: Normalize RI coverage & utilization rollup stats by sample count
**Background**
In querying statistics from a rollup table, the cost component treats data points for unique entities as equivalent to each other in terms of weight for averaging. This is incorrect given some data points may represent a larger number of samples. For example, in the scenario where VM_A is 100% covered and is up for a full hour (RI coverage latest will have 6 data points) and VM_B is 0% covered (100% on-demand) and is up for only a single topology broadcast. In the reserved_instance_coverage_by_hour rollup table, both VM_A and VM_B will have a single entry. In calculating an average for RI coverage, the cost component will calculate RI coverage as 50%, when the correct value would be ~86%.

At the impacted customer, this issue is exacerbated by the presence of a large number of ephemeral workloads running on-demand, which is significantly depressing the RI coverage calculated. The correct RI coverage is generally around 70-80% for the environment. However, daily rollups are calculated somewhere between 60-65% and month rollups are 5%.

**General Application**
The lack of weighting based on sample counts extends to other cost component statistics (e.g. cost, plan RI coverage & utilization). However, plan RI coverage & utilization is no longer shown in the UI and historical entity costs (through the workload cost breakdown widget) are generally analyzed for a specific entity. **XL Liability:** This issue most likely is not localized to the cost component. The history component, in providing general workload statistics, may not be correctly weighting downsampled averages based on sample count. This is a larger issue that will be addressed outside of this bug.

**Code Changes**
The proposed change is to no longer query for the average of rolled up RI coverage & utilization tables. Instead, we use the sample count (already maintained by the rollups for average) to calculate the total and then divide by the summed sample count. For the total, there is a hack put in place to normalize the value to Coupon hours (by dividing by 6, assuming 10 minute sample rate/topology broadcast).

**Testing Done**
1. Loaded the PROS diag onto an appliance (along with the RI coverage/utilization table dumps - thanks @greg_aligiannis). Verified the issue was reproducible.
2. Post fix, verified the daily and month values are consistent with the latest value (which does not suffer from a lack of sample weighting given each data point represents a single sample).