[OM-66970]: Allow voltron autocomplete in zsh
** Problem **

Starting with MacOS 10.15 (Catalina), Apple replaced bash with zsh as the default shell. The voltron autocompletion setup uses the compopt function which is not available in zsh. While it is possible to switch to bash, I would like it to work out of the box with zsh as well as bash.

** Solution **

Remove usage of the `compopt` function in the autocompletion setup. It seemed to be used to allow default shell completion behavior (autocomplete with files from the filesystem) for certain options. I changed the mechanism for allowing this by using arguments to the `complete` method as described in https://stackoverflow.com/a/22827323 to avoid use of compopt.
I also added a new voltron broadcast alias for voltron send_topology.
I also has a new command voltron get_topology for fetching the topology to stdout which you can then redirect to a file.

** Testing Done **

Ran on my new machine (running zsh) and autocompletion works for all options I tried. There is some extra setup required in ~/.zshrc along with these updates for the solution to work and I updated the voltron wiki page with this information. Also tried the same script on bash on my old laptop and it also works there.