[OM-65096]: Send keep alive progress message from actionscript probe every 15 seconds so the action does not timeout in topology processor
Intent
======
Prevent action timeout in topology processor when action takes a long time to execute in topology processor

Problem
=======
In an experiment that I run for executing large number of actions in actionscript, I noticed that around 10% of actions timed out in topology processor and failed

Diagnosis
=========
The default value for action timeout is 30 seconds in topology processor which means the action may timeout after 30 seconds if there is no activity in terms of progress with it from probe. In the actionscript probe, we send progress update every 60 seconds only if the has been an output from the script actionscript probe is running. This means the action may fail if it takes more than 30 secons.

Solution
========
There a polling thread (which polls every 60 seconds) that checks the output of ssh script and sends a progress update if there has been an output in the script. I changed the polling interval to 15 seconds and also introduced a keep-alive interval which 15 seconds. I updated the logic so we send the progress update message if there has been an output in ssh-script or the keep alive interval has passed. This means we send keep-alive message every 15 seconds which prevents action timeouts in topology processor.