[OM-66979]: Fix voltron hanging forever at startup
** Problem **

Voltron hangs forever at startup, never starting up.

** Solution **

The market component recently added a requirement to connect to the database at startup. However, when doing so they did not add the schema or password for the DB to voltron. This review fixes that.

** Testing Done **

Before, Voltron would hang forever at startup. After this fix, it is able to successfully start up.