[OM-66588] Create encryption master key
Now that we are reading in encryption keys from an external source (currently Kubernetes secrets and later Vault), we are depending on external management of multiple keys. The negative thing about this is that it is coupling the details of our microservices' security needs to the kubernetes secrets that we consume. Better might be to have a single "master" key that is read in through secrets, and manage the other keys internally to our system. The master key would be used to encrypt/decrypt the other keys. This could make operational changes like rotating the keys easier.