[OM-65686]: Include Container Pod Persistent Volumes in Plan Scope
** Problem **
Container pods with persistent volumes are unplaced in plans.

** Solution **
Container pods buy only storage amount (without a key) from their connected virtual volumes in kubeturbo. During plan scoping, we only pull in suppliers when they are connected through a commodity with a key. Aa a result, pod PVs are not pulled into plan scopes and the pods cannot be placed during market analysis.

The fix is to extend plan scoping semantics to cover the movable flag. When an entity sets movable to false, its only potential provider is its current provider. This commit updates the current logic to follow this logic by pulling in only the current provider regardless of keyed commodities when scoping.

Note that this update only affects the scoping done for non-cloud plans. Cloud plans such as migrate to cloud or optimize cloud use their own scoping logic which has not been updated.

** Testing Done **

* Ran optimize container plans with this change and ensured that container pods with PVs can now be placed.
* Updated unit tests for the new rules and ensured all existing unit tests still pass.
* Verified that the only probe which explicitly sets the movable flag is the kubernetes probe so that this change should be fairly low risk. Movable may also be set through things like settings and a poststitching operation for HyperV/VMM.