[OM-67216]: Prevent lowerBound > upperBound in ConsistentResizer
** Problem **

Sometimes consistent scaling actions generate resizes that exceed the configured max capacity.

As part of OM_63311, we introduced code in the ConsistentResizer to force the capacity lower bound to be a multiple of the capacity increment. We have a check that ensures capacity lower bound is less than or equal to capacity upper bound in the consistent resizer, but this check occurs BEFORE forcing the capacity lower bound to be a multiple of the capacity increment. For some values of the bounds and capacity increment, this change can cause the lower bound to become greater than the upper bound AFTER the check to ensure lower bound is less than or equal to the upper bound.

** Solution **

* After forcing capacity lower bound to be a multiple of capacity increment, prevent it from exceeding the upper bound.
* Note that this can cause capacity lower bound to no longer be a multiple of the capacity increment if the upper bound is not, but this is better than exceeding a limit set by the user.

** Testing Done **

* Ensured that the scenario in the bug no longer generates a resize exceeding the configured capacity upper bound.
* Added new unit test to cover new logic. Ensure all existing unit tests continue to pass.