[OM-67640]: Support automatic migration of admin JWT for upgrades
When upgrading from the legacy storage for secrets (using PVs) to the new approach using master keys, it is not easy to migrate the admin JWT. This file is used as a marker to indicate that the administrator user has been created. If this file is not migrated/recreated on an upgrade, then the user will be prompted to recreate the admin user following the upgrade--which is obviously not ideal.

The good news is that we can automatically recreate the JWT when we detect that an upgrade is happening (i.e., when we successfully import a legacy auth encryption key). It should be possible to do this strictly in the java code, without needed any additional kubernetes secrets to be created.

Note: Also fixed a bug where other components were re-using the 'auth' namespace in Consul for storing their keys, rather than using their own namespace.