[OM-64851]: Show action stats when we are scoping to a group of compute clusters
Intent
======
provide data for risks avoided widget when scoped to a group of clusters

Features Implemented
====================
Expand the group of cluster to individual clusters when asking for stats for a group of clusters. This was simply done by taking the same approach when we are dealing with a group of resource groups which is sending individual members of that groups of that resource group to the action orchestrator to get the stats rather than the group itself.

Testing Done
============
Added unit test for modified logic