[OM-68776]: Permit Voltron to show Swagger UI for external API
** Intent **

To be able to see the swagger UI for the external API in Voltron.

** Implementation **

I have not figured out how to successfully make external API requests from swagger to Voltron, but I have gotten the UI to display properly.

The problem was that in order to serve swagger resources properly, resources placed in two different directories by the build process need to be copied into a single directory. The docker plugin for the build pom does this. I introduced a Java class to become part of the Voltron workspace that does the same.

The new required env parameter API_COMPONENT_PATH should be the path on your local machine to the com.vmturbo.api.component codebase. Swagger files are copied from the build results for this component to the /swagger directory in the DATA_PATH.

** Remaining Work **
1. Figure out how to successfully make external API requests through Swagger UI to Voltron.
2. Figure out how to get the component-specific Swagger UI to be available in Voltron.

** Testing Done **

Verified that I can successfully view the Swagger UI in Voltron. The Swagger UI can be viewed at:
* http://localhost:8080/vmturbo/apidoc/index.html
Note that it is NOT visible at the alternate endpoint
* http://localhost:8080/swagger/index.html
I am not sure why the swagger-related endpoint doesn't work, but for Voltron's purposes having one working endpoit is sufficient.