[OM-68199]: Create the feature flag to enable local storage
Create the feature flag to enable local storage.

This flag must be set in the CR.yaml file, prior to running t8c.sh.

This task involves changes to the Operator as well as the deployment scripts. However, this task does not include removing Gluster in any way, simply ceasing to use it as the storage class.