[OM-68479]: Modify storage-restore script to skip gluster processing when using local storage
The storege-restore.sh script is called by turbo.service when the VM is restarted. Its purpose is to prevent the turbonomic pods from starting until gluster and heketi are both started and ready. There is no point in waiting when using local storage, so when this flag is set we skip the rest of the script.